--- server/src/main/java/io/spine/server/event/store/EEntityMatchesStreamQuery.java
@@ -1,49 +1,0 @@
-/*
- * Copyright 2019, TeamDev. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.server.event.store;
-
-import io.spine.core.Event;
-import io.spine.server.event.EventStreamQuery;
-import org.checkerframework.checker.nullness.qual.Nullable;
-
-import java.util.function.Predicate;
-
-/**
- * Tells if {@link EEntity} matches the passed query.
- */
-final class EEntityMatchesStreamQuery implements Predicate<EEntity> {
-
-    private final Predicate<Event> filter;
-
-    EEntityMatchesStreamQuery(EventStreamQuery query) {
-        this.filter = new MatchesStreamQuery(query);
-    }
-
-    @Override
-    public boolean test(@Nullable EEntity input) {
-        if (input == null) {
-            return false;
-        }
-        Event event = input.getState();
-        boolean result = filter.test(event);
-        return result;
-    }
-}
